# Add benchmark for Extract + small speed improvement