I'm working in that area, and will modify more things. Setting up a benchmark before going further

Bench results from that small change in the extraction func:
From `go test -benchmem -run=^$ -bench ^BenchmarkExtract$ github.com/grafana/tanka/pkg/process -count=10`

```
name               old time/op    new time/op    delta
Extract/regular-8    16.6µs ±20%    14.7µs ±19%  -11.47%  (p=0.019 n=10+10)
Extract/flat-8       7.09µs ±16%    6.18µs ±12%  -12.81%  (p=0.001 n=9+10)
Extract/deep-8       39.6µs ±13%    34.9µs ±13%  -11.92%  (p=0.000 n=9+9)
Extract/array-8      54.0µs ±27%    46.0µs ±13%  -14.68%  (p=0.006 n=10+9)
Extract/nil-8        17.6µs ±24%    13.8µs ± 6%  -21.48%  (p=0.000 n=10+8)

name               old alloc/op   new alloc/op   delta
Extract/regular-8    2.35kB ± 0%    2.22kB ± 0%   -5.44%  (p=0.000 n=10+10)
Extract/flat-8       1.06kB ± 0%    1.00kB ± 0%   -6.02%  (p=0.000 n=10+10)
Extract/deep-8       7.99kB ± 0%    7.74kB ± 0%   -3.20%  (p=0.000 n=10+10)
Extract/array-8      8.87kB ± 0%    8.55kB ± 0%   -3.61%  (p=0.000 n=10+10)
Extract/nil-8        2.38kB ± 0%    2.26kB ± 0%   -5.37%  (p=0.000 n=10+10)

name               old allocs/op  new allocs/op  delta
Extract/regular-8      45.0 ± 0%      41.0 ± 0%   -8.89%  (p=0.000 n=10+10)
Extract/flat-8         19.0 ± 0%      17.0 ± 0%  -10.53%  (p=0.000 n=10+10)
Extract/deep-8          130 ± 0%       122 ± 0%   -6.15%  (p=0.000 n=10+10)
Extract/array-8         153 ± 0%       143 ± 0%   -6.54%  (p=0.000 n=10+10)
Extract/nil-8          46.0 ± 0%      42.0 ± 0%   -8.70%  (p=0.000 n=10+10)
```